lab10 renew

+++ kernel/proc.h
@@ -91,7 +91,7 @@ struct vma{
   int flags;
   int fd;
   struct file *f;
-  uint64 offset;
+  int offset;
 };
 
 // Per-process state
@@ -118,6 +118,6 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
 
-  struct vma VMA[NVMA];
+     struct vma VMA[NVMA];  // Trace map
 };
